fix: desktop_interaction.py syntax errors + lint/formatting cleanup
- Fixed indentation and missing try/except blocks in desktop_interaction.py:
  - create_file(): added class indentation and exception handling
  - delete_file(): fixed structure with proper try/except
  - set_wallpaper(): added class indentation
- Fixed base_agent.py: corrected callback_address check to use task.payload.get()
- Applied black formatting (441 files) and isort import sorting
- Flake8 clean (0 errors)
- All tests passing: 149 API, 1340 core, 302 services, 488 multimodal, 328 garden, 1824 other AI tests

+++ apps/backend/src/ai/multimodal/quality_metrics.py
@@ -36,8 +36,8 @@ def ssim(
         sigma_b = np.var(channel_b)
         sigma_ab = np.mean((channel_a - mu_a) * (channel_b - mu_b))
 
-        numerator = (2 * mu_a * mu_b + c1) * (2 * sigma_ab + c2)
-        denominator = (mu_a**2 + mu_b**2 + c1) * (sigma_a + sigma_b + c2)
+        numerator = float((2 * mu_a * mu_b + c1) * (2 * sigma_ab + c2))
+        denominator = float((mu_a**2 + mu_b**2 + c1) * (sigma_a + sigma_b + c2))
         scores.append(float(numerator / max(denominator, 1e-8)))
     return float(np.mean(scores))
